Reminders: integer-day intervals only; remove plant from detail page
Reminders
- Drop sub-day interval options (10 min / 30 min / 1 hour) that were
  added for push testing. Production reminders are days-only now.
- Revert intervalDays GraphQL type from Float to Int (frontend mutation
  + backend schema). Backend min validation back to 1 + integer check.
- mapReminder defensively coerces any legacy fractional row to a whole
  number ≥ 1 so the Int! return type can't reject existing data.
- Time picker no longer needs the >= 1 day conditional — always shown.
- Drop the now-unused round6 helper and minuteCount/hourCount i18n keys.

Remove plant from detail page
- Add a "Remove plant" entry to the 3-dot menu (red, matches Plants
  list styling).
- Reuse the existing delete-confirmation modal pattern from
  plants.component.ts (same markup, same i18n keys: removeTitle,
  removeBody, keepIt, removing, common.remove). On confirm: calls
  plantService.remove and navigates back to /plants.

+++ backend/src/schema.ts
@@ -165,7 +165,7 @@ export const typeDefs = gql`
     id: String!
     plantId: String!
     actionType: ReminderActionType!
-    intervalDays: Float!
+    intervalDays: Int!
     notifyTime: String
     nextDueAt: String!
     snoozedUntil: String
@@ -230,7 +230,7 @@ export const typeDefs = gql`
     adminCreateUser(email: String!, password: String!, role: String!, tier: String!): User!
     adminUpdateUser(userId: String!, email: String, role: String, tier: String): User!
     adminDeleteUser(userId: String!): Boolean!
-    setPlantReminder(plantId: String!, actionType: ReminderActionType!, intervalDays: Float!, enabled: Boolean!, notifyTime: String): PlantReminder!
+    setPlantReminder(plantId: String!, actionType: ReminderActionType!, intervalDays: Int!, enabled: Boolean!, notifyTime: String): PlantReminder!
     disableAllReminders: Int!
     removePlantReminder(id: String!): Boolean!
     snoozeReminder(id: String!, hours: Int): PlantReminder!
